Move Core functions to a new component named MorseTranslator
- Create /MorseTranslator with index.tsx and MorseTranslator.tsx
- Move components into MorseTranslator.tsx
- Import component in App.tsx
